pyproject.toml
[tool.poetry]
name = "trainingproj"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.38"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
pandas = "^2.2.3"
beautifulsoup4 = "^4.13.3"
openpyxl = "^3.1.5"
xlrd = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
